+++ Frameworks/MASearchSuggestions/MASearchSuggestions/MAProperties.swift
@@ -5,8 +5,12 @@
 //  Created by Ashwin Paudel on 2021-12-30.
 //
 
+import AppKit
 import Foundation
 
 public enum MAProperties {
-    static let cornerRadius: CGFloat = 7
+    static var cornerRadius: CGFloat = 7
+    static var highlightColor: NSColor = .selectedContentBackgroundColor
+    static var backgroundColor: NSColor = .labelColor
+    static var textColor: NSColor = .labelColor
 }